HEEDLS-201 Use UAT values for hardcoded SCORM content
As the SCORM urls in the db refer to content hosted in the live application,
running them with customisation values from the test db won't work.

The UAT values won't work when running locally or on test, but should
work when deployed to UAT, so that should be enough for a proof of
concept.