HEEDLS-201 Change test old system url
The old system url was set to live on test. This was
partially a random choice as integration with the old
system won't work on test anyway. However having it point
to UAT will be slightly more useful. Because some SCORM
content might work, if the ids are the same in the test
and uat db, in which case it's preferrabel for the tracking
to be updated in the UAT db.