HEEDLS-198 Update login count and duration (#38)
Added an method for updating LoginCount and Duration in Progress; this required a new method for obtaining ProgressID from Progress using CandidateID and CustomisationID which was also added. This update will be ran whenever a learning menu is opened.

Ran and passed all unit tests then manually tested by changing the stored value then opening the page on Firefox.