HEEDLS-209 Back end for initial course page
Add Section class, containing basic data about a candidates progress
through a specific section.

Amend GetCourseContent class to contain a list of sections, and the get
method in the CourseContentService to collect all of this.

Modify LearningMenuController to use the candidate ID to find a course
overview, so details about progress through each section can be given.

Amend tests to reflect new structure of models, and add some to cover
the edge cases of getting a course.