HEEDLS-172 Add brand logo to couse page (#47)
Add brand logos to course page if centre logo is not available.

Add a LogoService to manage finding logos for a page, which replaces
GetCentreLogo in the CentresService. This will find the best logo based
on the given criteria, for a centreId and customisationId. If a
customisationId is not given the centre logo is still returned if it
exists.

Replace the CentreLogo model with the more general Logo model for both
centre and brand logos. Move the resizing of the logos from the service
to CSS (because that requires height and width fields for the logos,
which are not recorded in the Brand table unlike the Centre logo).

Pass the customisationId from the LearningMenu view up to the _Layout
and into the LogoViewComponent, who uses it to call the LogoService to
get the logo.

Refactor and simplify the logo ViewModel and View, and control its size
using a CSS class in layout.css instead of in these classes.

Remove GetCentreLogo tests, and add tests for the LogoService.