Create progress entry on first launch (#49)
Add a new method to check if a given entry exists in the progress table and executes the uspCreateProgressRecord_V3 stored procedure if the entry doesn't exist.

The stored procedure has a few checks and doesn't insert the entry if any fail. This will be logged but the program will still try to run GetProgress method so this has been kept as QueryFirstOrDefault to prevent crashing if the entry wasn't added.

Ran and passed all unit tests and manually tested on Firefox, "Current Activities" will show newly enrolled courses.