Add default values to TutorialContent query
When a candidate is not enrolled on a course but queries a tutorial in
it, return the tutorial with progress values set to a default (eg.
TimeSpent = 0).

Adjust the TutorialContentService tests to reflect this.

Automatically enrolling a user after returning a tutorial if they are
not already enrolled is part of the LearningMenuController already.