Fix too many sections on initial menu
Inner join against the CustomisationsTutorials to make sure we only
consider tutorials selected for this course customisation.

Fix a percentage complete counting bug, which when counting the number
of tutorials that need to be completed it considered all potential
tutorials instead of just the ones selected for this course.

Fixing this bug solved the main issue of the ticket.