DLSV2-56 Implements Competency Group Add/Edit
View with viewmodel and associated controller and services. Also tweaks service loading of list for My Framework to avoid duplicates without introducing GROUP BY clause.
When editing a Framework Competency Group, the service checks if the Competency Group is used in other frameworks and creates a new one if so.