HEEDLS-302 Add single section front end (#177)
Amend initial menu action to redirect to the section page if the
course has only one section.

Add completion data to section view model, to be shown if it is the
only section in the course, and to flag to record whether this is the
only section or not. Also hide the next button in this case.

Amend tests to reflect these changes to the initial menu action, and
these additions to the section content view model.