HEEDLS-313 Add single tutorial frontend (#183)
Amend section action to redirect to the tutorial page if the course
has only one tutorial.

Add completion data to tutorial view model, to be shown if it is the
only tutorial in the course, and flags to record whether this is the
only section, and only item in the section. Also hide the next
button in this case.

Amend tests to reflect these changes to the section action, and these
additions to the tutorial view model.