Add single post learning assessment front end
Amend section action to redirect to the post learning assessment page
if the course only has a post learning assessment.

Add completion data to post learning assessment view model, to be shown
if it is the item in the course, and flags to record whether this is
the only section, and only item in the section.

Hide the next button if this is the only item in the section.

Modify the mobile-only back breadcrumb to have a more helpful name,
depending on if there are other sections in the course or not.