Add single diagnostic assessment front end
Amend section action to redirect to the diagnostic assessment page if
the course only has a diagnostic assessment.

Add completion data to diagnostic assessment view model, to be shown
if it is the only item in the course, and flags to record whether this
is the only section, and only item in the section.

Hide the next button if this is the only item in the section.

Modify the mobile-only back breadcrumb to have a more helpful name,
depending on if there are other sections in the course or not.