HEEDLS-332 Change tutorial next link to button (#220)
Only show this next button if the tutorial has been completed.

Tweak some CSS to deal with the long row of buttons. We can't
use the NHS width classes unfortunately because we need to 
switch to the vertical list of buttons when the screen is narrower
than ~1030px but the widest NHS width class (large-desktop) is
only 990px wide.